map_app/requirements.txt
aiohttp==3.9.1
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
click==8.1.7
ecdsa==0.18.0
fastapi==0.109.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
pyasn1==0.5.1
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.25.0
yarl==1.9.4